+++ save-common/src/commonMain/kotlin/org/cqfn/save/core/plugin/TempDirException.kt
@@ -1,0 +1,6 @@
+package org.cqfn.save.core.plugin
+
+/**
+ * An [Exception] that can be thrown when create/delete temp dir
+ */
+class TempDirException(message: String) : Exception(message)
